% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_md.R
\name{paleta_md}
\alias{paleta_md}
\title{Paleta de colores institucional de MetaDocencia}
\format{
Vector de caracteres con códigos hex.
}
\usage{
paleta_md
}
\description{
Vector con los cinco colores institucionales en orden recomendado
para uso en gráficos categóricos.
}
\examples{
\dontrun{
paleta_md
}

}
