% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_md.R
\name{scale_color_md}
\alias{scale_color_md}
\alias{scale_colour_md}
\title{Escala de color discreta con paleta MetaDocencia}
\usage{
scale_color_md(...)

scale_colour_md(...)
}
\arguments{
\item{...}{Argumentos adicionales pasados a \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Para usar con \verb{ggplot2::scale_color_*} y \verb{scale_fill_*}.
}
\examples{

library(ggplot2)

ggplot(mpg, aes(displ, hwy, color = class)) +
  geom_point(size = 3) +
  scale_color_md() +
  theme_md()


}
