% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_md.R
\name{scale_fill_md}
\alias{scale_fill_md}
\title{Escala de relleno discreta con paleta MetaDocencia}
\usage{
scale_fill_md(...)
}
\arguments{
\item{...}{Argumentos adicionales pasados a \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
Escala de relleno discreta con paleta MetaDocencia
}
\examples{
library(ggplot2)

ggplot(mpg, aes(class, fill = class)) +
  geom_bar() +
  scale_fill_md() +
  theme_md()


}
